requirements.txt
Flask==3.0.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.26.2
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
pdf2image==1.16.3
Pillow==10.1.0
APScheduler==3.10.4
openpyxl==3.1.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0
google-generativeai==0.3.2
transformers==4.35.2
# Use a compatible torch release available on pip for this environment
torch>=2.2.0
sentence-transformers==2.2.2
openai==1.3.7

